LPS-55528 Move DB Store class to the module
