+++ portal-impl/src/portal.properties
@@ -8990,7 +8990,6 @@
     # document library server will use this to persist documents.
     #
     #dl.store.impl=com.liferay.portlet.documentlibrary.store.AdvancedFileSystemStore
-    #dl.store.impl=com.liferay.portlet.documentlibrary.store.DBStore
     dl.store.impl=com.liferay.portlet.documentlibrary.store.FileSystemStore
 
     #
